+++ GoodBuyUS/src/main/java/dev_final_team10/GoodBuyUS/controller/api/SSEController.java
@@ -1,4 +1,4 @@
-package dev_final_team10.GoodBuyUS.controller;
+package dev_final_team10.GoodBuyUS.controller.api;
 
 import dev_final_team10.GoodBuyUS.service.SSEService;
 import org.springframework.http.MediaType;
